Refactored Code to make ease of use
- Added buttons, textinput, text area to shared components
- refactored comment code to a component instead of page so custom hook is being called from the component.
