+++ pages/_app.js
@@ -1,15 +1,14 @@
 import "../styles/theme.css";
 import "../styles/globals.css";
 import "tailwindcss/tailwind.css";
-
-import MainLayout from "../components/Layout";
+import MainLayout from "../shared/components/Layout";
 
 function MyApp({ Component, pageProps }) {
-    return(
-    <MainLayout>
-         <Component {...pageProps} />
-    </MainLayout>
-    )
+    return (
+        <MainLayout>
+            <Component {...pageProps} />
+        </MainLayout>
+    );
 }
 
 export default MyApp;
